pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corpus-mlx"
version = "0.2.1"
description = "CorePulse × MLX Stable Diffusion: prompt injection, masking, regional control (src layout)"
authors = [{name="CorePulse Contributors"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "mlx>=0.5.0",
    "mlx-lm>=0.0.6",
]

[project.optional-dependencies]
stable-diffusion = [
    "Pillow>=9.0.0",
    "huggingface-hub>=0.19.0",
    "safetensors>=0.4.0",
    "tqdm",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["corpus_mlx"]
